Fixed various compiler warnings from g++ -Wall
Tidied some constructor member-orderings to avoid need for #pragma.
Removed selected unused variables to avoid "#pragma unused".
Defensively initialized various variables to avoid uninitialized warnings.
Patched miscellaneous boolean ORs masquerading as logical ORs.
(Many warnings remain.)

+++ dev/ext/lib/Procframe/GeneralConstraint.cpp
@@ -613,7 +613,7 @@ namespace gpstk
          varNum++;
       }
       
-      if(varNum!=convertMat.rows() || varNum!=convertMat.cols())
+      if(varNum != (int)convertMat.rows() || varNum != (int)convertMat.cols())
       {
          Exception e("The size of input doesn't match.");
          GPSTK_THROW(e);
